Separates resource release from task end notification
The Coordinator now releases all resources at the same time. Only then
does it notify actors that their tasks have finished.

The problem this solves:

Assume a task A of simulation P is attached to resource R.
Assume a task B of a different simulation Q is attached to another
resource S.

Once B is finished, Q will check on the availability of
resource R. Assume that if R is available, Q will run a task X, else
it will run a task Y.

If A and B finish at the same time, R will be released, so Q should
(always) run X.

Before this commit, if B finishes first and is handled by
`handleCEvent`, Q will proceed to check R which has not been detached
yet. It will think R is still busy and it will go to Y.

This commit makes sure all resources from finished tasks (R, S) are
detached first before any of the simulations (P, Q) are notified that
their respective tasks (A, B) have finished.

This way, when Q finds out B has finished, it will check on R which
will already be detached irrespective of the order of the events.