Implements lookahead via scheduler-sim messages
Lookahead is implemented through the scheduler asking the simulation
about what comes after every task that is scheduled. This is done by
creating the `LookaheadScheduler` and a `Lookahead` trait in the
simulation.

The `LookaheadScheduler` works almost exactly like `DefaultScheduler`
but it also appends new future tasks to the tasks list, and it receives
these tasks by asking the simulation using the
`Simulation.TasksAfterThis` message after a task is added.

If a simulation is created with the `Lookahead` trait, it can add
information to a `lookaheadMap` which is used to reply to the scheduler.
The `lookaheadMap` is used to reply to the scheduler with a  list of
`Task` objects which will be created in the future when the task with
the UUID given in the `TasksAfterThis` message completes.

I believe it is bad to use `TaskGenerator.create()` in the Simulation,
but this was the easiest way to give the scheduler what it needs. In the
future maybe some other way of communicating 'future tasks' should exist
so that we don't need to create task objects. However, to do so I think
a new type of task class would need to be made which is lightweight, can
be used by the scheduler and simulation, and can be created from a
`Task` object. Alternatively maybe it is possible to 'pre-load' all
tasks into the coordinator and have the coordinator reply to the
scheduler's questions.

The `LookaheadTest` class is temporary. I used it for testing the new
scheduler and chose to include it since it might come in handy in the
future.