Implements support for Many-To-One future tasks
Had to completely re-do the lookaheadMap idea, now future events are
described in a bunch of `Set`s instead of a `Map`, this lets me use a
function to describe when a task should start (said function returns -1
if the corresponding task cannot start yet).

Added a few methods to make interface nicer: `add1To1Lookahead`,
`add1ToManyLookahead`, `addManyTo1Lookahead`.

Added DummySim3 for testing an AND flow-style condition.

Had to re-arange and add a lot of random things to get everything to
work.