Adds `LookaheadObj` Class
objects of this new class are passed to the scheduler so that the
scheduler does not have to message the simulation every time it needs to
ask about a new future task. Such objects are defined in the simulation
and passed innitialy via the coordinator (hence the introduction of one
new coordinator message). The scheduler stores these lookahead objects
in a map and these objects can be updated at any time by the
simulation- this is typically done whenever a task completes, but this
could possibly be optimised in the future.

Modifies the existing tests to work with the new class.