Implements lookahead strucutre
Changes the way the lookahead object works by splitting the
functionality into a lookahead structure which contains immutable feilds
only, and a wrapper in the scheduler which interfaces with said
structure.

This change introduces the `LookaheadStructure` trait, and defines
`LookaheadObj` as a case class of this trait. The trait and case class
provide methods for adding and removing entries in the structure as well
as adding two structures together.

The way the scheduler works has been changed to work with this
structure. The `tasksAfterThis` method which was previously within the
lookahead object has been moved to the scheduler, and the way it works
has changed slightly. The old `lookaheadNextIter` method has been
removed completely as with this approach it is no longer
necessary. The way `inProgressFutureTasks` (previously named `r`) are
found has also changed. The entire `getNextTasks` method is very
different, one crucial detail is the usage of
`lookaheadObjects.values.reduce((a,b)=>a+b)` which combines all of the
lookahead objects into one great big object which is then used in the
`findNextTasks` function. The `findNextTasks` function has been
modified too, it now takes `scheduledThisIter` and
`lookaheadSetThisIter` as inputs and it modifies these inputs as it
calls itself recursively.

Unit tests have been updated. The old way of testing scheduler doesn't
work, I commented it all out for now but new tests will need to be
written.