adds basic tests and modifies parseFlow
changed parseFlow to use a LookaheadStrcture paramater so that it no
longer relies on the existance of a `var lookahead`.

Updated EmptyStructure so that EmptyStructure + (something) returns an
actual structure such that it is possible to grow an empty structure.

Made simple tests to test parseFlow- some of these fail at the moment,
I'm debating if the behaviour is appropriate or if it should be changed
to match the tests expected output.