Fixes issues caused by previous merge
Fixes flows and FlowTest so that they work now
Adds some minor things to taskGenerator that were missing because I put
them on a wrong branch by accident.
Updates some other tests and Simulation to remove errors that arrise as
a result of the changes to how taskGenerators are defined.