Moves completed set from lookahead to scheduler
`LookaheadStructure`s no longer require the `complete` and `uncomplete`
methods. The tracking of completed tasks is now done in the scheduler.

This affected the way `LookaheadStructure.-` works, and for now the
scheduler removes some additional tasks before it begins scheduling
tasks to make up for the change, but in the future it would be
preferable to remove these tasks in the simulation. Maybe a
restructuring of simulation classes is needed...