Adds futureTask tag to `Task`s
Allows scheduler to mark tasks as future tasks
Fixes the issue where the scheduler was able to return future tasks even
though the coordinator never asked about them. Now a task that is
created for the purpose of lookahead will not be returned because it is
marked as a future task.