Removes isFutrueTask, adds completed set to sim
Removes isFutureTask from `Task`s, instead the scheduler now returns the
intersection of tasks and the result to guarnatee that future tasks are
not returned.

Simulation tracks completed tasks for lookahead, and removes the
appropriate entries from the structure locally instead of relying on the
scheduler to update the structures.

Also includes a few small tweaks which address various TODOs from
wednesday's meeting.