WIP working with JBPM to get tasks.
Trying for months now to get the tasks from a bpmn file in order such
that we can simulate them. This commit contains some essentials for
running the jBPM library and some test bpmn files in the resources
folder. The build file has been edited to include all the necessary
dependencies (and there are a lot of them).

As of writing it is my understanding that working with abstract tasks
and script tasks is not possible. Human tasks are the last hope, but
they have proven difficult to work with due to requiring users which are
members of various groups. I have yet to make even one working human
model.