Fixes bug in `Flow`s where tasks would be removed after callback
Now that we do not have asynchronicity any more, the callback triggers
`Simulation.ack` which tells the `Coordinator` to continue the time.
This mean that the task was removed from the `tasks` map at the end of
the simulation, which in turn made subsequent callbacks (such as the
one by `And` which relies on the `tasks` map being up to date) to misbehave.