Adds SimSubrunHandler, MetaMetrics
MetaMetrics (probably a temporary name) are used to store certain KPI
information using the `BasicStats` structure.

`SimSubrunHandler` is an alternative to the original `SimOutputHandler`:
A certain number of subruns is specified and Simulation Aggregate data
will be divided up between the appropriate number of SimMetricsHandlers,
effectively using multiple SimMetricsAggregators for a single simulation
whereby each aggregator represents a single subrun.

In the future the first subrun might be excluded from calculating
confidence intervals to give the effect of a warmup period.

This commit includes some small modifications to MainTest as a result of
ongoing testing.