Deletes MockSimulationActors tests
The rationale is that this is overkill for the functionality that is
being tested. Simulation behaviour is already covered by the
coordinator tests. This is only meant to test the interface between
the actor and the underlying simulation. This should be simpler.

Given the low priority given to the Akka implementation at the moment,
it may be a while until other tests will be added for this.