Updates `Publisher.subscribe` to a `Resource` + first unit test
With `use` we returned an IO of a Stream to be used later. However,
once that terminates (generating the expected Stream), `use`
completes and releases the resource (unsubscribes).

Instead, we now use `flatMap` and return the generated Stream as a
Resource. Any use of the subsription must happen within the call of
`use` though.

Next I need to see how to scale to multiple subs and add more tests.