Ensures `Simulation` termination always happens last
See unit test: "correctly work with a case starting at the time limit"

If a case starts at the same time as the time limit, then `stop()` is
called before the case is notified to start (via `run()`). This
results in `CaseRef.stop()` being called before `run()`.

`Simulation.handleDiscreteEvent` is more complicated now, with an
extra boolean parameter, but oh well, such is life. This is not meant
to be exposed anywhere so it's fine.